+++ src/pki/federal/__init__.py
@@ -1,5 +1,6 @@
 """DoD and Federal PKI provider definitions built on pki-core."""
 
+from .algorithms import SP800_78_ALGORITHM_POLICY
 from .providers import (
     BUILTIN_PROVIDERS,
     CAC_PROVIDER,
@@ -14,6 +15,7 @@ __all__ = [
     "CAC_PROVIDER",
     "ECA_PROVIDER",
     "PIV_PROVIDER",
+    "SP800_78_ALGORITHM_POLICY",
     "default_registry",
     "full_registry",
 ]